requirements.txt
arrow==1.2.3
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
binaryornot==0.4.4
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
cookiecutter==2.1.1
Django==4.1.5
django-browser-reload==1.6.0
django-tailwind==3.4.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
flake8==5.0.4
flake8-django==1.1.5
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytest==7.2.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-slugify==8.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
urllib3==1.26.14
